Feature/join dinner

+++ app/Http/Controllers/DinnerController.php
@@ -3,8 +3,13 @@
 namespace App\Http\Controllers;
 
 use App\Models\Dinner;
+use App\Models\DinnerInvite;
+use Carbon\Carbon;
+use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class DinnerController extends Controller {
     /**
@@ -13,7 +18,11 @@ class DinnerController extends Controller {
      * @return Response
      */
     public function index() {
-        return response( Dinner::all() );
+        $dinners = ( new Dinner )->whereDate( 'start_time', '>', Carbon::now() )->get();
+        if ( $dinners->isNotEmpty() ) {
+            return response( $dinners );
+        }
+        throw new NotFoundHttpException( 'No dinners found.' );
     }
 
     /**
@@ -31,8 +40,8 @@ class DinnerController extends Controller {
             'start_time'  => 'after:now|date',
             'end_time'    => 'after:start_time|date'
         ] );
-
-        return response( ( new Dinner )->create( $request->all() ) );
+        $userId    = auth( 'api' )->user()->id;
+        return response( ( new Dinner )->create( array_merge( $request->all(), [ 'user_id' => $userId ] ) ), 201 );
     }
 
     /**
@@ -43,7 +52,20 @@ class DinnerController extends Controller {
      * @return Response
      */
     public function show( $id ) {
-        return response( ( new Dinner )->find( $id ) );
+        $dinner = $this->findDinner( $id );
+        if ( $dinner != null ) {
+            return response( $dinner );
+        }
+        throw new NotFoundHttpException( 'Dinner not found.' );
+    }
+
+    /**
+     * @param $id
+     *
+     * @return Dinner|Dinner[]|Collection|Model|null
+     */
+    private function findDinner( $id ) {
+        return ( new Dinner )->find( $id );
     }
 
     /**
@@ -55,10 +77,13 @@ class DinnerController extends Controller {
      * @return Response
      */
     public function update( Request $request, $id ) {
-        $dinner = ( new Dinner )->find( $id );
-        $dinner->update( $request->all() );
+        $dinner = $this->findDinner( $id );
+        if ( $dinner != null ) {
+            $dinner->update( $request->all() );
 
-        return response( $dinner );
+            return response( $dinner );
+        }
+        throw new NotFoundHttpException( 'Dinner not found.' );
     }
 
     /**
@@ -69,6 +94,11 @@ class DinnerController extends Controller {
      * @return Response
      */
     public function destroy( $id ) {
-        return response( Dinner::destroy( $id ) );
+        $dinner = $this->findDinner( $id );
+        if ( $dinner != null ) {
+            return response( Dinner::destroy( $id ) );
+        }
+        throw new NotFoundHttpException( 'Dinner not found.' );
     }
+
 }
